% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{adapt_grdev_args}
\alias{adapt_grdev_args}
\title{Adapt the args in the call for a particular graphical device}
\usage{
adapt_grdev_args(gr_dev, passed_args)
}
\arguments{
\item{gr_dev}{the device type}

\item{passed_args}{a list of passed args}
}
\description{
Adapt the args in the call for a particular graphical device
}
