% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{adapt_grdev_file}
\alias{adapt_grdev_file}
\title{Adapt the file and filename parameters}
\usage{
adapt_grdev_file(gr_dev, passed_args)
}
\arguments{
\item{gr_dev}{the device type}

\item{passed_args}{a list of passed args}
}
\description{
Adapt the file and filename parameters
}
