% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{aplt}
\alias{aplt}
\title{Create a reasonably sized figure given extents and constraints}
\usage{
aplt(
  xlim,
  ylim,
  code,
  ...,
  .pretty_axes = TRUE,
  .max_width = 10,
  .max_height = 8,
  .max_distortion = 2,
  .dpi = 96,
  .envir = parent.frame()
)
}
\arguments{
\item{xlim}{the x extents}

\item{ylim}{the y extents}

\item{code}{the code to run after calling the graphics device}

\item{...}{graphical parameters}

\item{.pretty_axes}{use pretty to update specified xlim and ylim}

\item{.max_width}{max figure width in inches}

\item{.max_height}{max figure height in inches}

\item{.max_distortion}{the maximum allowable distortion ratio}

\item{.dpi}{the dots per inch}

\item{.envir}{the execution environment in which to process code block}
}
\description{
Create a reasonably sized figure given extents and constraints
}
