% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{app_factory}
\alias{app_factory}
\title{Create a CLI app object with custom behavior.}
\usage{
app_factory()
}
\value{
a cli app object
}
\description{
This monkey patches the behavior of the "inline" method so that it may
use a custom transformer, specified as an option.
}
