% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{appender_jupyter}
\alias{appender_jupyter}
\title{An appender for jupyter (with stderr)}
\usage{
appender_jupyter(lines)
}
\description{
In jupyter, stderr will show up on the console, and messages will show up
in the notebook.  We emit messages only in the jupyter context to avoid
double entries.
}
\keyword{internal}
