% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject.R
\name{bind_call}
\alias{bind_call}
\title{Bind an injected function}
\usage{
bind_call(ref, cmd_call, ...)
}
\arguments{
\item{ref}{the reference function to bind}

\item{cmd_call}{a cmd call}
}
\description{
Bind an injected function
}
