% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{build_vignette}
\alias{build_vignette}
\title{Build a vignette from a jupyter notebook}
\usage{
build_vignette(ipynb_name)
}
\arguments{
\item{ipynb_name}{the ipython notebook name}
}
\description{
Jupyter notebooks should be stored in ./vignettes directory.
}
