% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{capture_cli_message}
\alias{capture_cli_message}
\title{Capture text from cli message}
\usage{
capture_cli_message(msg)
}
\arguments{
\item{msg}{a cli_message}
}
\description{
Capture text from cli message
}
