% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-cli.R
\name{captured_cli_opts}
\alias{captured_cli_opts}
\title{logger appropriate cli options}
\usage{
captured_cli_opts()
}
\value{
a list of options
}
\description{
logger appropriate cli options
}
\keyword{internal}
