% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{cfplot}
\alias{cfplot}
\title{Provide a plot context that is independent of publishing context}
\usage{
cfplot(
  code,
  antialias = NULL,
  bg = NULL,
  canvas = NULL,
  colormodel = NULL,
  compress = NULL,
  display = NULL,
  encoding = NULL,
  family = NULL,
  fg = NULL,
  file = NULL,
  filename = NULL,
  fillOddEven = NULL,
  fonts = NULL,
  gamma = NULL,
  height = 7,
  onefile = NULL,
  pagecentre = NULL,
  paper = NULL,
  pointsize = 12,
  res = 96,
  symbolfamily = NULL,
  title = NULL,
  type = NULL,
  units = "in",
  useDingbats = NULL,
  useKerning = NULL,
  version = NULL,
  width = 7,
  xpos = NULL,
  ypos = NULL,
  ...,
  .envir = parent.frame()
)
}
\arguments{
\item{height}{height}

\item{res}{resolution}

\item{units}{units of width and height}

\item{width}{width}

\item{...}{passed to dev_args}

\item{.envir}{environment to execute code block}
}
\description{
Provide a plot context that is independent of publishing context
}
