% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_wrapped.R
\name{cli_bullets}
\alias{cli_bullets}
\title{Wrapped version of cli::cli_bullets}
\usage{
cli_bullets(text, id = NULL, class = NULL, .envir = parent.frame())
}
\arguments{
\item{text}{Character vector of items. See details below on how names
are interpreted.}

\item{id}{Optional id of the \code{div.bullets} element, can be used in themes.}

\item{class}{Optional additional class(es) for the \code{div.bullets} element.}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
It is often useful to print out a list of items, tasks a function or
package performs, or a list of notes.
}
