% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_wrapped.R
\name{cli_format}
\alias{cli_format}
\title{Wrapped version of cli::cli_format}
\usage{
cli_format(x, style = NULL, ...)
}
\arguments{
\item{x}{The object to format.}

\item{style}{List of formatting options, see the individual methods
for the style options they support.}

\item{...}{Additional arguments for methods.}
}
\description{
This function can be used directly, or via the \verb{{.val ...}} inline
style. \verb{{.val {expr}}} calls \code{cli_format()} automatically on the value
of \code{expr}, before styling and collapsing it.
}
