% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_wrapped.R
\name{cli_list_themes}
\alias{cli_list_themes}
\title{Wrapped version of cli::cli_list_themes}
\usage{
cli_list_themes()
}
\value{
A list of data frames with the active themes.
Each data frame row is a style that applies to selected CLI tree nodes.
Each data frame has columns:
\itemize{
\item \code{selector}: The original CSS-like selector string. See \link[cli]{themes}.
\item \code{parsed}: The parsed selector, as used by cli for matching to nodes.
\item \code{style}: The original style.
\item \code{cnt}: The id of the container the style is currently applied to, or
\code{NA} if the style is not used.
}
}
\description{
If there is no active app, then it calls \code{\link[cli:start_app]{start_app()}}.
}
