% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-handler.R
\name{cli_nullity_handler}
\alias{cli_nullity_handler}
\title{replace NULLs with visible values}
\usage{
cli_nullity_handler(msg)
}
\arguments{
\item{msg}{a cli-message object}
}
\description{
replace NULLs with visible values
}
