% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_wrapped.R
\name{cli_progress_along}
\alias{cli_progress_along}
\title{Wrapped version of cli::cli_progress_along}
\usage{
cli_progress_along(
  x,
  name = NULL,
  total = length(x),
  ...,
  .envir = parent.frame()
)
}
\arguments{
\item{x}{Sequence to add the progress bar to.}

\item{name}{Name of the progress bar, a label, passed to
\code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}

\item{total}{Passed to \code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}

\item{...}{Passed to \code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}

\item{.envir}{Passed to \code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}
}
\value{
An index vector from 1 to \code{length(x)} that triggers progress
updates as you iterate over it.
}
\description{
Note that this function is currently experimental!

Use \code{cli_progress_along()} in a mapping function or in a for loop, to add a
progress bar. It uses \code{\link[cli:cli_progress_bar]{cli_progress_bar()}} internally.
}
