% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_wrapped.R
\name{cli_progress_message}
\alias{cli_progress_message}
\title{Wrapped version of cli::cli_progress_message}
\usage{
cli_progress_message(
  msg,
  current = TRUE,
  .auto_close = TRUE,
  .envir = parent.frame(),
  ...
)
}
\arguments{
\item{msg}{Message to show. It may contain glue substitution and cli
styling. It can be updated via \code{\link[cli:cli_progress_update]{cli_progress_update()}}, as usual.}

\item{current}{Passed to \code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}

\item{.auto_close}{Passed to \code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}

\item{.envir}{Passed to \code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}

\item{...}{Passed to \code{\link[cli:cli_progress_bar]{cli_progress_bar()}}.}
}
\value{
The id of the new progress bar.
}
\description{
This is a simplified progress bar, a single (dynamic)
message, without progress units.
}
