% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_wrapped.R
\name{cli_progress_num}
\alias{cli_progress_num}
\title{Wrapped version of cli::cli_progress_num}
\usage{
cli_progress_num()
}
\value{
\code{cli_progress_num()} returns an integer scalar.

`cli_progress_cleanup() does not return anything.
}
\description{
Progress bar utility functions.
}
