% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_wrapped.R
\name{cli_progress_output}
\alias{cli_progress_output}
\title{Wrapped version of cli::cli_progress_output}
\usage{
cli_progress_output(text, id = NULL, .envir = parent.frame())
}
\arguments{
\item{text}{Text to output. It is formatted via \code{\link[cli:cli_text]{cli_text()}}.}

\item{id}{Progress bar id. The default is the current progress bar.}

\item{.envir}{Environment to use for glue interpolation of \code{text}.}
}
\value{
\code{TRUE}, always.
}
\description{
The text is calculated via \code{\link[cli:cli_text]{cli_text()}}, so all cli features can be
used here, including progress variables.
}
