% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_wrapped.R
\name{cli_progress_styles}
\alias{cli_progress_styles}
\title{Wrapped version of cli::cli_progress_styles}
\usage{
cli_progress_styles()
}
\value{
A named list with sublists containing elements
\code{complete}, \code{incomplete} and potentially \code{current}.
}
\description{
The following options are used to select a style:
\itemize{
\item \code{cli_progress_bar_style}
\item \code{cli_progress_bar_style_ascii}
\item \code{cli_progress_bar_style_unicode}
}
}
