% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrap.R
\name{cli_wrap_safe}
\alias{cli_wrap_safe}
\title{Wrap an exported function, safely}
\usage{
cli_wrap_safe(name, level)
}
\arguments{
\item{name}{the object (function) name}

\item{level}{the logger level}
}
\value{
a wrapped function
}
\description{
Wrap an exported function, safely
}
\keyword{internal}
