% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{construct_rmd_file}
\alias{construct_rmd_file}
\title{Internal:  construct an R markdown file}
\usage{
construct_rmd_file(cfg, content, rmd_pth)
}
\arguments{
\item{cfg}{a config}

\item{content}{the json extracted content of a jupyter notebook}

\item{rmd_pth}{the r markdown path}
}
\description{
Internal:  construct an R markdown file
}
\keyword{internal}
