% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-postprocess.R
\name{decode_knitr}
\alias{decode_knitr}
\title{Internal: decode the start of an encoded knitr chunk}
\usage{
decode_knitr(enc, dict)
}
\arguments{
\item{dict}{an environment to use as a dictionary}
}
\description{
Internal: decode the start of an encoded knitr chunk
}
