% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generic.R
\name{deprecated}
\alias{deprecated}
\title{Raise an error for deprecated functions}
\usage{
deprecated(is_terminal = TRUE)
}
\arguments{
\item{is_terminal}{if TRUE, abort; if FALSE, warn}
}
\description{
Raise an error for deprecated functions
}
