% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{display_highlight}
\alias{display_highlight}
\title{Highlight user provided code}
\usage{
display_highlight(content, header = NULL)
}
\arguments{
\item{content}{code to be highlighted, a character}

\item{header}{a header}
}
\description{
Highlight user provided code
}
