% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{display_source}
\alias{display_source}
\title{Highlighted R source files and inject HTML into document}
\usage{
display_source(path)
}
\arguments{
\item{path}{a path to a source file or directory}
}
\description{
Highlighted R source files and inject HTML into document
}
