% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{display_text}
\alias{display_text}
\title{"as.character", applied uniformly across context}
\usage{
display_text(obj)
}
\arguments{
\item{obj}{an R object}
}
\description{
"as.character", applied uniformly across context
}
