% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdev.R
\name{.with_grdev}
\alias{.with_grdev}
\title{Internal.  Execute a code block in a graphics context.}
\usage{
.with_grdev(
  grdev_name,
  .expr,
  dev_args = list(),
  par_opts = list(),
  .envir = parent.frame()
)
}
\arguments{
\item{grdev_name}{the name of the graphics object}

\item{.expr}{the expression to evaluate}

\item{dev_args}{args to pass to the device}

\item{par_opts}{options to pass to par}

\item{.envir}{environment to execute code block}
}
\value{
invisible image data
}
\description{
Internal.  Execute a code block in a graphics context.
}
