% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monkey-patch.R
\name{.with_monkey_patch}
\alias{.with_monkey_patch}
\title{Internal.  Temporarily monkey patch a namespaced method}
\usage{
.with_monkey_patch(fqn, wrapper, expr, .envir = parent.frame())
}
\arguments{
\item{fqn}{the fully qualified name, e.g. "rmarkdown::html_document"}

\item{wrapper}{a wrapper with signature function(cmd, args, func)}

\item{expr}{an expression}

\item{.envir}{the environment in which to evaluate expr}
}
\description{
Internal.  Temporarily monkey patch a namespaced method
}
