% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-postprocess.R
\name{encode_knitr}
\alias{encode_knitr}
\title{Internal: encode the start of a knitr chunk as a hash}
\usage{
encode_knitr(l, dict)
}
\arguments{
\item{l}{the line to encode}

\item{dict}{an environment to use as a dictionary}
}
\description{
Internal: encode the start of a knitr chunk as a hash
}
