% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-env.R
\name{env_merge}
\alias{env_merge}
\title{Merge environments}
\usage{
env_merge(dest, src)
}
\arguments{
\item{dest}{the destination environment}

\item{src}{the source environment}
}
\description{
Merge src into dest.  Src will overwrite elements in dest.
}
