% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-env.R
\name{env_stack}
\alias{env_stack}
\title{Return the environment stack}
\usage{
env_stack(e, last = getOption("pkgette.env_stack_last"))
}
\arguments{
\item{e}{the environment; or function}

\item{last}{passed to the underlying call to env_parents}
}
\value{
a list of environments
}
\description{
Utility function that wraps env_parents.  However, it always includes the current environment
and will handle function environments.
}
