% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-env.R
\name{env_strip}
\alias{env_strip}
\title{Return the names in the environment stack}
\usage{
env_strip(f)
}
\arguments{
\item{f}{an the environment; or function}
}
\value{
a character vector of the environment names in the stack
}
\description{
Note that this removes hash suffixes.
}
