% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-env.R
\name{extend_env_search}
\alias{extend_env_search}
\title{Extend an environment from search path.}
\usage{
extend_env_search(pkg_names, envir)
}
\arguments{
\item{pkg_names}{}

\item{envir}{the "base" environment}
}
\value{
a new environment with the specified lineage
}
\description{
This should work for names that are already in the search path.
}
