% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{extract_jpy_cells}
\alias{extract_jpy_cells}
\title{Internal:  extract jupyter notebook cells}
\usage{
extract_jpy_cells(content)
}
\arguments{
\item{content}{the json extracted content of a jupyter notebook}
}
\description{
Internal:  extract jupyter notebook cells
}
