% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{extract_yaml_cells}
\alias{extract_yaml_cells}
\title{Internal:  extract tagged yaml cells from ipython notebook}
\usage{
extract_yaml_cells(content)
}
\arguments{
\item{content}{the json extracted content of a jupyter notebook}
}
\value{
the coalesced YAML from "yaml" tagged cells
}
\description{
Internal:  extract tagged yaml cells from ipython notebook
}
