% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{get_fileset}
\alias{get_fileset}
\title{Get fileset from a file name or directory}
\usage{
get_fileset(path)
}
\arguments{
\item{path}{a file path, a directory path; if relative, assummed to be wrt the package root}
}
\description{
Get fileset from a file name or directory
}
