% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{get_grdev_close_device}
\alias{get_grdev_close_device}
\title{Return the default close_device value}
\usage{
get_grdev_close_device(gr_dev = c("png", "pdf", "x11"))
}
\arguments{
\item{gr_dev}{the device type}

\item{passed_args}{a list of passed args}
}
\description{
Return the default close_device value
}
