% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{get_grdev_ext}
\alias{get_grdev_ext}
\title{Get the grdev extension}
\usage{
get_grdev_ext(gr_dev = c("png", "pdf", "x11"))
}
\arguments{
\item{gr_dev}{the device type}
}
\description{
Get the grdev extension
}
