% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{get_grdev_op}
\alias{get_grdev_op}
\title{Get the grdev operator}
\usage{
get_grdev_op(gr_dev = c("png", "pdf", "x11"))
}
\arguments{
\item{gr_dev}{the device type}
}
\description{
Get the grdev operator
}
