% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdev.R
\name{grdev_adapt}
\alias{grdev_adapt}
\title{For the specified grdev, select an appropriate subset of an argset}
\usage{
grdev_adapt(grdev, argset)
}
\arguments{
\item{grdev}{a graphics object acquired from grdev_get}

\item{argset}{a passed argument set}
}
\description{
This also incorporates default arguments
}
