% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdev.R
\name{grdev_get}
\alias{grdev_get}
\title{Get a graphics object by its name.}
\usage{
grdev_get(grdev_name = c("png", "pdf", "x11"))
}
\arguments{
\item{grdev_name}{the name of the graphics object}
}
\description{
Get a graphics object by its name.
}
