% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdev.R
\name{grdev_resolve}
\alias{grdev_resolve}
\title{Resolve any ambiguities in file / filename args}
\usage{
grdev_resolve(grdev, argset)
}
\arguments{
\item{grdev}{a graphics object acquired from grdev_get}

\item{argset}{a passed argument set}
}
\description{
Resolve any ambiguities in file / filename args
}
\keyword{internal}
