% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generic.R
\name{hash_trim}
\alias{hash_trim}
\title{Trim a hash value}
\usage{
hash_trim(val)
}
\arguments{
\item{val}{the (hash) value to trim}
}
\description{
Set the "mty.hash_label_length" option to change the default (identity)
}
