% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{highlight}
\alias{highlight}
\title{A refactored / shared code highlight function}
\usage{
highlight(content, header = NULL)
}
\arguments{
\item{content}{the json extracted from a jupyter notebook}

\item{header}{a header}
}
\description{
A refactored / shared code highlight function
}
