% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{highlight_code_method}
\alias{highlight_code_method}
\title{Assemble codes to be highlighted}
\usage{
highlight_code_method(args, method = c("fileset", "raw"))
}
\arguments{
\item{args}{a fileset; or raw source code}

\item{method}{either "fileset" or "raw"}
}
\description{
Assemble codes to be highlighted
}
