% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{highlight_source}
\alias{highlight_source}
\title{Highlight R source files and inject HTML into document}
\usage{
highlight_source(args, theme = "seashell", method = c("fileset", "raw"))
}
\arguments{
\item{args}{depends on method}

\item{theme}{a knitr css theme}

\item{method}{the method to use}
}
\description{
Highlight R source files and inject HTML into document
}
\keyword{internal}
