% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{hook_metayer}
\alias{hook_metayer}
\title{The metayer/knitr hook used for preprocessing / postprocessing}
\usage{
hook_metayer(before, options, envir, name, ...)
}
\arguments{
\item{before}{a boolean}

\item{options}{the current chunk options}

\item{envir}{the environment in which our code will be executed}

\item{name}{name associated with the hook, e.g. "metayer"}

\item{...}{to match knitr hook signature}
}
\description{
Refer to \href{https://yihui.org/knitr/hooks}{knitr documentation}
}
