% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{initialize_logging}
\alias{initialize_logging}
\title{Setup logging}
\usage{
initialize_logging(
  r_config_active = Sys.getenv("R_CONFIG_ACTIVE", "default"),
  home = fs::path_home(),
  max_bytes = 1000000L,
  max_files = 7L,
  create_directory = TRUE
)
}
\arguments{
\item{r_config_active}{the config section to use}

\item{home}{the user's home directory}

\item{max_bytes}{the max_bytes parameter passed to logger::appender_file}

\item{max_files}{the max_files parameter passed to logger::appender_file}

\item{create_directory}{a boolean, TRUE to create the directory}
}
\description{
Setup logging
}
