% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{initialize_vignette}
\alias{initialize_vignette}
\title{Initialize for common vignette settings}
\usage{
initialize_vignette(highlight_theme = "seashell")
}
\arguments{
\item{highlight_theme}{the code highlight theme}

\item{disable_cli}{disable cli, a boolean}
}
\description{
Called from onLoad
}
