% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_inject.R
\name{inj_body_cli}
\alias{inj_body_cli}
\title{Inject code body}
\usage{
inj_body_cli(cmd, level, .caller_env = caller_env())
}
\arguments{
\item{cmd}{a call, the cli operation that we're wrapping}

\item{level}{the logger level to associated with the cmd}
}
\description{
wrap_factory uses the body of this function as a building block.  It
does this using 'substitute', replacing the cmd and level parameters
with their known expressions.
}
\details{
Note, the awkward substitute construction:

local_env <- new_environment(list(cmd = cmd))
injection <- lapply(
body(inj_body_cli),
function(s) do.call(substitute, list(s, local_env))
) \%>\%
as.call()
}
