% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_inject.R
\name{inj_get_cmd}
\alias{inj_get_cmd}
\title{Programmatically construct an external call}
\usage{
inj_get_cmd(pkg, name, func = NULL)
}
\arguments{
\item{pkg}{a package name, e.g., "cli"}

\item{name}{a function in the package, e.g., "cli_text"}

\item{func}{the function, if available}
}
\value{
a call that contains the requested command with args
}
\description{
Programmatically construct an external call
}
