% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-publish.R
\name{ipynb_yaml_extract}
\alias{ipynb_yaml_extract}
\title{extract YAML from a notebook}
\usage{
ipynb_yaml_extract(ipynb_in)
}
\arguments{
\item{ipynb_in}{an ipynb file}
}
\description{
Returns a nested list comprised of merged YAML from raw notebook cells.
}
\keyword{internal}
