% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{is_header_cell}
\alias{is_header_cell}
\title{Internal: is the cell tagged as "yaml"}
\usage{
is_header_cell(c)
}
\arguments{
\item{c}{the cell to test}
}
\description{
Internal: is the cell tagged as "yaml"
}
