% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-postprocess.R
\name{is_knitr}
\alias{is_knitr}
\title{Internal:  is this the start of a knitr chunk?}
\usage{
is_knitr(l)
}
\arguments{
\item{l}{the line to test}
}
\description{
Internal:  is this the start of a knitr chunk?
}
