% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-publish.R
\name{is_knitr_chunk_start}
\alias{is_knitr_chunk_start}
\title{Internal:  is this the start of a knitr chunk?}
\usage{
is_knitr_chunk_start(l)
}
\arguments{
\item{l}{the line to test}
}
\description{
Internal:  is this the start of a knitr chunk?
}
