% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-postprocess.R
\name{is_knitr_enc}
\alias{is_knitr_enc}
\title{Internal:  is this the start of an encoded knitr chunk?}
\usage{
is_knitr_enc(l)
}
\arguments{
\item{l}{the line to test}
}
\description{
Internal:  is this the start of an encoded knitr chunk?
}
