% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-publish.R
\name{jpy_wrap}
\alias{jpy_wrap}
\title{quick wrap a vignette for jupyter config}
\usage{
jpy_wrap(code, .envir = parent.frame())
}
\arguments{
\item{code}{client code}

\item{.envir}{env in which to evaluate client code}
}
\description{
quick wrap a vignette for jupyter config
}
\keyword{internal}
