% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{list_coalesce}
\alias{list_coalesce}
\title{Internal: coalesce lists}
\usage{
list_coalesce(agg, y)
}
\arguments{
\item{agg}{the base case list}

\item{y}{the new list to coalesce into the base case list}
}
\value{
a coalesced list
}
\description{
Internal: coalesce lists
}
\keyword{internal}
