% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_abort}
\alias{log_abort}
\title{Produce a NULL-aware "abort" message}
\usage{
log_abort(message, ..., .class = NULL, .parent = NULL, .envir = parent.frame())
}
\arguments{
\item{message}{the user provided message}

\item{...}{to be passed through to cli::cli_bullets}

\item{.class}{an error class}

\item{.parent}{a parent exception, if warranted}

\item{.envir}{the environment in which to evaluate the underlying glue function}
}
\description{
Produce a NULL-aware "abort" message
}
