% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_alert_info}
\alias{log_alert_info}
\title{Generate a NULL-aware cli_alert_info}
\usage{
log_alert_info(
  text,
  id = NULL,
  class = NULL,
  wrap = FALSE,
  .envir = parent.frame()
)
}
\arguments{
\item{text}{the user provided text}

\item{id}{id, to be passed through to cli::cli_alert_*}

\item{class}{class, to be passed through to cli::cli_alert_*}

\item{wrap}{wrap, to be passed through to cli::cli_alert_*}

\item{.envir}{the environment in which to evaluate the underlying glue function}
}
\description{
Generate a NULL-aware cli_alert_info
}
