% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_debug}
\alias{log_debug}
\title{Log a message with given log level}
\usage{
log_debug(...)
}
\arguments{
\item{...}{R objects that can be converted to a character vector via the active message formatter function}
}
\value{
Invisible \code{list} of \code{logger} objects. See \code{\link[logger]{logger}} for more details on the format/
}
\description{
Log a message with given log level
}
\examples{
\dontrun{
log_level(INFO, 'hi there')
log_info('hi there')

## output omitted
log_debug('hi there')

## lower threshold and retry
log_threshold(TRACE)
log_debug('hi there')

## multiple lines
log_info('ok {1:3} + {1:3} = {2*(1:3)}')

log_layout(layout_json())
log_info('ok {1:3} + {1:3} = {2*(1:3)}')

## note for the JSON output, glue is not automatically applied
log_info(glue::glue('ok {1:3} + {1:3} = {2*(1:3)}'))
}
}
\seealso{
\code{\link[logger]{logger}}
}
