% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_inform}
\alias{log_inform}
\title{Produce a NULL-aware "inform" message}
\usage{
log_inform(message, ..., .envir = parent.frame())
}
\arguments{
\item{message}{the user provided message}

\item{...}{to be passed through to cli::cli_bullets}

\item{.envir}{the environment in which to evaluate the underlying glue function}
}
\description{
Produce a NULL-aware "inform" message
}
