% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-test.R
\name{log_level_mock_factory}
\alias{log_level_mock_factory}
\title{A factory for mocked log_level calls}
\usage{
log_level_mock_factory(filename)
}
\arguments{
\item{filename}{the redirected output file name}
}
\description{
Use this to redirect logger::log_level to a (temporary) file
}
