% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{log_wrapper}
\alias{log_wrapper}
\title{Common code shared by the other log_* functions}
\usage{
log_wrapper(
  handler,
  threshold,
  message,
  ...,
  collapse = TRUE,
  strip_newline = FALSE,
  .envir = parent.frame()
)
}
\arguments{
\item{handler}{handler, to pass to cli::cli_fmt}

\item{threshold}{a logging emit threshold}

\item{message}{the user provided message}

\item{...}{to be passed through to cli::cli_fmt}

\item{collapse}{to be passed through to cli::cli_fmt}

\item{strip_newline}{to be passed through to cli::cli_fmt}

\item{.envir}{the environment in which to evaluate the underlying glue function}
}
\description{
This wraps cli::cli_fmt with options that enable the NULL-aware logging
}
