% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-handler.R
\name{logged_cli_handler}
\alias{logged_cli_handler}
\title{Redirect CLI messages to logger}
\usage{
logged_cli_handler(msg)
}
\arguments{
\item{msg}{a cli_message}
}
\description{
Redirect CLI messages to logger
}
