% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/docs.R
\name{metayer-config}
\alias{metayer-config}
\title{metayer environment variables and options}
\description{
metayer environment variables and options
}
\section{User facing configuration}{

\subsection{config.yml}{

Package configuration is specified in config.yml, which is designed to mimic the behavior of the \href{https://rstudio.github.io/config/reference/index.html}{config} package.  As such, there are currently three sections: default, testing, and jupyter.  Each section may adapt the following subsections: logger, options.

When metayer is loaded, it automatically sets the options and sets up the loggers.

"!optenv" is a YAML tag that will be evaluated by R.  It is useful for specifying R symbols in a configuration context, e.g., \verb{!optenv logger::appender_console}.  Note, there are security implications here.  Think before using this and don't be malicious.
\subsection{\code{mty.cli_null}}{

For cli methods, NULL values will be replaced by this value, e.g. "\if{html}{\out{<null>}}".
}

}

\subsection{jupyter notebook YAML}{

We introduce a YAML tag, "!with_env" that does a glue injection with environmental variables.  This is useful for directing pandoc to a configurable location.
}
}

