% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{metayer_handler}
\alias{metayer_handler}
\title{A CLI message handler}
\usage{
metayer_handler(msg)
}
\arguments{
\item{msg}{a cli_message}
}
\description{
This is a variant of cli::cli_server_default_safe
}
