% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-handler.R
\name{msg_rewrite_nulls}
\alias{msg_rewrite_nulls}
\title{Rewrite NULL values in a cli message, optionally}
\usage{
msg_rewrite_nulls(msg)
}
\arguments{
\item{msg}{a cli message}
}
\value{
the message type
}
\description{
If the mty.cli_null is NULL, this is a no-op.
}
\keyword{internal}
