% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generic.R
\name{mty_abort}
\alias{mty_abort}
\title{A logged abort}
\usage{
mty_abort(message, .envir = parent.frame())
}
\arguments{
\item{message}{a glue-able string}

\item{.envir}{the environment in which to evaluate the message}
}
\description{
A logged abort
}
