% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generic.R
\name{mty_salted_hash}
\alias{mty_salted_hash}
\title{Create a predictable identifier sequence}
\usage{
mty_salted_hash(salt = NULL)
}
\description{
This is probably most useful when testing, as one can set the corresponding entry
in the config.yml to get reproducible results.
}
