% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generic.R
\name{mty_uuid}
\alias{mty_uuid}
\title{get a uuid}
\usage{
mty_uuid(...)
}
\arguments{
\item{...}{pass through parameters}
}
\description{
This can be adapted to produce seeded results by specifying the uuid.generator option, e.g.
options(uuid.generator = test_mty_uuid)
}
