% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generic.R
\name{not_yet_implemented}
\alias{not_yet_implemented}
\title{Raise an error for not yet implemented functions}
\usage{
not_yet_implemented(is_terminal = TRUE)
}
\arguments{
\item{is_terminal}{if TRUE, abort; if FALSE, warn}
}
\description{
Raise an error for not yet implemented functions
}
