% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{null_aware_transformer}
\alias{null_aware_transformer}
\title{A NULL aware alternative to cli:::inline_transformer}
\usage{
null_aware_transformer(code, envir)
}
\arguments{
\item{code}{The text inside the "..." glue substitution}

\item{envir}{Environment with the data to perform the styling.}
}
\value{
the output of cli:::inline_transformer after a "NULL" substitution
}
\description{
A NULL aware alternative to cli:::inline_transformer
}
