% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{opt_plotbox}
\alias{opt_plotbox}
\title{Return optimized figure extents}
\usage{
opt_plotbox(W, H, dx, dy, distortion_ratio = 2)
}
\arguments{
\item{W}{the maximum figure width}

\item{H}{the maximum figure height}

\item{dx}{the diff of xlim}

\item{dy}{the diff of ylim}

\item{distortion_ratio}{the maximum allowable distortion ratio, i.e.: (w / h) / (dx / dy)}
}
\description{
Ref: https://predict-epfl.github.io/piqp/
}
