% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{pcd}
\alias{pcd}
\title{A simple data.frame used in vignettes}
\format{
\subsection{\code{pcd}}{

A data frame with 10 rows and 2 columns:
\describe{
\item{x}{an independent variables}
\item{y}{a dependent variable}
}
}
}
\usage{
pcd
}
\description{
A simple data.frame used in vignettes
}
\keyword{internal}
