% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pimg.R
\name{pimg_as_html_data}
\alias{pimg_as_html_data}
\title{Wrap png_data in an HTML img wrapper}
\usage{
pimg_as_html_data(png_data, alt = NULL)
}
\arguments{
\item{png_data}{raw png data}
}
\value{
an IMG element
}
\description{
Wrap png_data in an HTML img wrapper
}
\keyword{internal}
