% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{plot_context}
\alias{plot_context}
\title{Provide a plot context that is independent of publishing context}
\usage{
plot_context(
  code,
  ...,
  .width = 7,
  .height = 7,
  .units = "in",
  .res = 96,
  .pointsize = 12,
  .envir = parent.frame()
)
}
\arguments{
\item{code}{the plotting code to evaluatie}

\item{...}{graphical parameters}

\item{.width}{figure width, in inches}

\item{.height}{figure height, in inches}

\item{.units}{the units}

\item{.res}{the dots per inch}

\item{.pointsize}{pointsize}

\item{.envir}{the environment in which to evaluate code}
}
\description{
Provide a plot context that is independent of publishing context
}
