% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{png_wrap}
\alias{png_wrap}
\title{Wrap png_data in an HTML img wrapper}
\usage{
png_wrap(png_data)
}
\arguments{
\item{png_data}{raw png data}
}
\value{
an IMG element
}
\description{
Wrap png_data in an HTML img wrapper
}
