% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-postprocess.R
\name{postprocess_rmd}
\alias{postprocess_rmd}
\title{postprocess output of rmarkdown::convert_ipynb}
\usage{
postprocess_rmd(input, output = xfun::with_ext(input, "Rmd"))
}
\arguments{
\item{input}{input file, a jupyter notebooks}

\item{output}{output file, an Rmd file}
}
\description{
This allows jupyter notebooks to inject YAML that overrides the usual convert_ipynb
defaults, namely, titles like:
}
\details{
"An R Markdown document converted from "vignettes/examples.ipynb"
}
