% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-publish.R
\name{preknit}
\alias{preknit}
\title{Knit an rmd file before handing it off to a downstream publishing stage.}
\usage{
preknit(
  rmd_in,
  md_out = NULL,
  root_dir = NULL,
  requested_format = c("html", "pdf")
)
}
\arguments{
\item{rmd_in}{input Rmd file}

\item{md_out}{output md file}

\item{root_dir}{knitr root.dir}

\item{requested_format}{currently, "html" or "pdf"}
}
\description{
Knit an rmd file before handing it off to a downstream publishing stage.
}
