% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-publish.R
\name{prep_articles}
\alias{prep_articles}
\title{Prepare notebooks for downstream processing}
\usage{
prep_articles(..., regexp = "*Rmd")
}
\arguments{
\item{...}{the path for Rmd files, passed to here::here}

\item{regexp}{a regular expression for filtering files}
}
\description{
Like a Makefile, specify Rmd files as targets, and processes ipynb files to create them.
}
