% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-publish.R
\name{process_rmd}
\alias{process_rmd}
\title{postprocess output of rmarkdown::convert_ipynb}
\usage{
process_rmd(rmd_in, rmd_out, conf = list())
}
\arguments{
\item{conf}{a nested list that will be converted into a YAML header}

\item{input}{input file, an Rmd file as prepared by rmarkdown::convert_ipynb}

\item{output}{output file, an Rmd file}
}
\description{
This is part of a multi-step process that converts jupyter notebooks into viable knitr
Rmd files.
}
\details{
In particular, it introduces a per-cell unique identifier and allows the caller to
replace any existing YAML in the input.  The latter is useful because the existing
implementaton of rmarkdown::convert_ipynb produces an Rmd file with a fixed title
and no mechanism for injecting additional, or modifying existing, YAML configuration.
}
