% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{pub_rmd_to_html}
\alias{pub_rmd_to_html}
\title{Convert an rmd to an html}
\usage{
pub_rmd_to_html(rmd_in, html_out = NULL)
}
\arguments{
\item{rmd_in}{the rmd filename}

\item{html_out}{the pdf filename}
}
\description{
Convert an rmd to an html
}
