% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{pub_rmd_to_md}
\alias{pub_rmd_to_md}
\title{knit an rmarkdown file}
\usage{
pub_rmd_to_md(rmd_in, md_out = NULL, dev = NULL)
}
\arguments{
\item{rmd_in}{input file, an Rmd file as prepared by rmarkdown::convert_ipynb}

\item{md_out}{output file, an Rmd file}

\item{dev}{for knitr.chunk.dev}
}
\description{
Knit an rmarkdown file into a markdown file suitable for pandoc processing.
}
