% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{pub_rmd_to_pdf}
\alias{pub_rmd_to_pdf}
\title{Convert an rmd to a pdf}
\usage{
pub_rmd_to_pdf(rmd, .pdf)
}
\arguments{
\item{rmd}{the rmd filename}

\item{.pdf}{the pdf filename}
}
\description{
Convert an rmd to a pdf
}
