% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubcontext.R
\name{pubcontext}
\alias{pubcontext}
\title{dispatch contextualized code}
\usage{
pubcontext(
  jupyter_code = NULL,
  knitr_code = NULL,
  rstudio_code = NULL,
  interactive_code = NULL,
  non_interactive_code = NULL,
  raise = FALSE,
  .envir = parent.frame(),
  .local_envir = parent.frame()
)
}
\arguments{
\item{jupyter_code}{jupyter client code}

\item{knitr_code}{knitr client code}

\item{rstudio_code}{rstudio client code}

\item{interactive_code}{interactive script client code}

\item{non_interactive_code}{non-interactive script client code}

\item{raise}{boolean, if TRUE, raise errors}

\item{.envir}{environment in which to evaluate code block}

\item{.local_envir}{environment in which to process logs}
}
\description{
Pubcontext dispatches contextualized code.
}
