% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{publish_context}
\alias{publish_context}
\title{Output the html in the correct publishing context}
\usage{
publish_context(html)
}
\arguments{
\item{html}{some html to publish}
}
\description{
Output the html in the correct publishing context
}
\keyword{internal}
