% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapped-factory.R
\name{remap_symb}
\alias{remap_symb}
\title{Remap symbols in a list}
\usage{
remap_symb(args, envir = parent.frame())
}
\arguments{
\item{args}{the passed args, a list of symb = symb}
}
\value{
a list where the value symbols have been evaluated
}
\description{
Remap symbols in a list
}
