% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{render_rmd_file}
\alias{render_rmd_file}
\title{Internal:  render an HTML from R markdown file}
\usage{
render_rmd_file(rmd_pth)
}
\arguments{
\item{rmd_pth}{the rmd file name}
}
\description{
Internal:  render an HTML from R markdown file
}
