% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{reset_options_from_conf}
\alias{reset_options_from_conf}
\title{Reset options from config.yml}
\usage{
reset_options_from_conf(
  r_config_active = Sys.getenv("R_CONFIG_ACTIVE", "default")
)
}
\arguments{
\item{r_config_active}{the config section to use}
}
\description{
Reset options from config.yml
}
\keyword{internal}
