% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{reset_pander_from_conf}
\alias{reset_pander_from_conf}
\title{Reset pander options from config.yml}
\usage{
reset_pander_from_conf()
}
\description{
Reset pander options from config.yml
}
\keyword{internal}
