% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-publish.R
\docType{data}
\name{rexp}
\alias{rexp}
\title{regular expressions for encode / decode functions}
\format{
An object of class \code{character} of length 1.
}
\usage{
rexp
}
\description{
regular expressions for encode / decode functions
}
\keyword{datasets}
