% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-generic.R
\name{rm.all}
\alias{rm.all}
\title{Utility function that resets the global environment.}
\usage{
rm.all(exclusions = c())
}
\arguments{
\item{exclusions}{a list of object names to be excluded from the removal.}
}
\description{
Utility function that resets the global environment.
}
