% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{set_envvar_defaults}
\alias{set_envvar_defaults}
\title{Set environmental variables to default values if necessary}
\usage{
set_envvar_defaults()
}
\description{
Set environmental variables to default values if necessary
}
\keyword{internal}
