% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{set_highlight_theme}
\alias{set_highlight_theme}
\title{Set the highlight theme; output HTML}
\usage{
set_highlight_theme(theme)
}
\arguments{
\item{theme}{a published theme name}
}
\description{
Choose from published themes:  knitr::knit_theme$get()
preview:  http://animation.r-forge.r-project.org/knitr/
}
\details{
This needs to operate in two modes.  One is within Jupyter; the other is
when building the html.
}
