% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/initialize.R
\name{setup_logging}
\alias{setup_logging}
\title{Setup logging}
\usage{
setup_logging(
  logfile_template,
  home = fs::path_home(),
  user = Sys.info()[["login"]],
  max_bytes = 1000000L,
  max_files = 7L,
  create_directory = TRUE,
  tee = TRUE
)
}
\arguments{
\item{logfile_template}{a glue template}

\item{home}{the user's home directory}

\item{user}{the user's login name}

\item{max_bytes}{the max_bytes parameter passed to logger::appender_file}

\item{max_files}{the max_files parameter passed to logger::appender_file}

\item{create_directory}{a boolean, TRUE to create the directory}

\item{tee}{a boolean, TRUE sends to file and stdout; FALSE, only file}
}
\description{
Setup logging
}
