% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{source_footer_template}
\alias{source_footer_template}
\title{Internal: a template for source files}
\usage{
source_footer_template()
}
\description{
Internal: a template for source files
}
\keyword{internal}
