% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage_env}
\alias{storage_env}
\title{retrieve a storage environment}
\usage{
storage_env(..., .store = ".storage", .envir = rlang::global_env())
}
\arguments{
\item{...}{an indexing collection of character objects}

\item{.store}{the name of the storage}

\item{.envir}{the environment into which to bind the storage data structure}
}
\value{
the specified environment
}
\description{
retrieve a storage environment
}
