% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/storage.R
\name{storage_purge}
\alias{storage_purge}
\title{Purge the storage}
\usage{
storage_purge(store = ".storage", envir = rlang::global_env())
}
\arguments{
\item{envir}{the environment into which to purge the storage_env data structure}

\item{.store}{the name of the storage}
}
\description{
Purge the storage
}
