% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sureplot.R
\name{sure_plot}
\alias{sure_plot}
\title{Create uniform plots across publishing contexts}
\usage{
sure_plot(
  code,
  width = getOption("mty.fig.width", 7),
  height = getOption("mty.fig.height", 7),
  res = getOption("mty.fig.res", 96),
  units = getOption("mty.fig.units", "in"),
  file = tempfile(),
  ...,
  par_opts = list(),
  .envir = parent.frame()
)
}
\arguments{
\item{code}{client code}

\item{width}{width}

\item{height}{height}

\item{res}{resolution dpi}

\item{units}{units of width and height}

\item{file}{a file, mapped to filename, if the device expects it}

\item{...}{parameters passed to device args}

\item{par_opts}{graphics pararameters}

\item{.envir}{the environment in which to evaluate the client code}
}
\description{
Create uniform plots across publishing contexts
}
