% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{sure_source}
\alias{sure_source}
\title{Highlight a source file or the source files in a directory}
\usage{
sure_source(path, theme = "seashell", raw = FALSE)
}
\arguments{
\item{path}{a path}

\item{theme}{a knitr theme}

\item{raw}{if TRUE, 'path' will be interpreted as lines of code}
}
\description{
Highlight a source file or the source files in a directory
}
