% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{sure_text}
\alias{sure_text}
\title{Wraps utils::capture.output for uniformity across publishing context.}
\usage{
sure_text(obj)
}
\arguments{
\item{obj}{an R object}
}
\description{
Wraps utils::capture.output for uniformity across publishing context.
}
