% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-test.R
\name{test_filter}
\alias{test_filter}
\title{A less motivational test}
\usage{
test_filter(test_name)
}
\arguments{
\item{test_name}{the canonical stub name, e.g., 'logger'}
}
\description{
A less motivational test
}
