% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-test.R
\name{test_mty_uuid}
\alias{test_mty_uuid}
\title{Create a predictable identifier sequence}
\usage{
test_mty_uuid(seed = NULL)
}
\arguments{
\item{seed}{a seed for the hash function}
}
\description{
This is probably most useful when testing, as one can set the corresponding entry
in the config.yml to get reproducible results.
}
