% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-test.R
\name{test_sanitize}
\alias{test_sanitize}
\title{Basic test sanitation}
\usage{
test_sanitize(exclusions = c(), envir = parent.frame())
}
\arguments{
\item{exclusions}{a list of object names to exclude from rm.all}

\item{envir}{a local environment}
}
\description{
This:
\itemize{
\item runs any deferred tasks that may be pending
\item calls devtools::load_all if not testing
\item resets loggers
\item purges the storage
\item calls rm.all with noted exclusions
\item sets up a deferred storage purge on exit
}
}
