% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-cli.R
\name{wcli_cnd}
\alias{wcli_cnd}
\title{Map a wrapped cli condition to its source}
\usage{
wcli_cnd(wcnd)
}
\arguments{
\item{cnd}{a wrapped cli message}
}
\description{
Calling catch_cnd on a wrapped cli condition will return a verbatim cli message
which masks the source condition.  This method extracts that source condition.
}
