% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{with_dev}
\alias{with_dev}
\title{Provide a default context}
\usage{
with_dev(
  code,
  ...,
  .width = 7,
  .height = 7,
  .pointsize = 12,
  .envir = parent.frame()
)
}
\description{
Provide a default context
}
