% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdev.R
\name{with_grdev}
\alias{with_grdev}
\title{Execute a code block in a graphics context.}
\usage{
with_grdev(
  grdev_name,
  code,
  width = getOption("mty.fig.width", 7),
  height = getOption("mty.fig.height", 7),
  res = getOption("mty.fig.res", 96),
  units = getOption("mty.fig.units", "in"),
  ...,
  par_opts = list(),
  .envir = parent.frame()
)
}
\arguments{
\item{grdev_name}{the name of the graphics object}

\item{width}{width}

\item{height}{height}

\item{res}{resolution}

\item{units}{units of width and height}

\item{...}{passed to dev_args}

\item{par_opts}{options to pass to par}

\item{.envir}{environment to execute code block}
}
\description{
Execute a code block in a graphics context.
}
