% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{with_logger}
\alias{with_logger}
\title{wrap client code with logging functionality}
\usage{
with_logger(code, .local_envir = parent.frame(), level = NULL)
}
\arguments{
\item{code}{client code}

\item{.local_envir}{environment to evaluate client code}

\item{level}{a logger level; may override condition message defaults}
}
\description{
wrap client code with logging functionality
}
