% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-test.R
\name{with_message_buf}
\alias{with_message_buf}
\title{Capture messages in a bytestream}
\usage{
with_message_buf(code, nbytes = 1000, strip_newline = TRUE)
}
\arguments{
\item{code}{client code to execute}

\item{nbytes}{the size of the bytearray}
}
\value{
the captured output
}
\description{
Capture messages in a bytestream
}
