% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/monkey-patch.R
\name{with_monkey_patch}
\alias{with_monkey_patch}
\title{Temporarily monkey patch a namespaced method}
\usage{
with_monkey_patch(fqn, wrapper, code, .envir = parent.frame())
}
\arguments{
\item{fqn}{the fully qualified name, e.g. "rmarkdown::html_document"}

\item{wrapper}{a wrapper with signature function(cmd, args, func)}

\item{code}{user code}

\item{.envir}{the environment in which to evaluate expr}
}
\description{
Temporarily monkey patch a namespaced method
}
