% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander.R
\name{with_pander}
\alias{with_pander}
\title{Contextualized pander}
\usage{
with_pander(code, .envir = parent.frame())
}
\arguments{
\item{code}{client code}

\item{.envir}{environment in which to evalutes client code}
}
\description{
Contextualized pander
}
