% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{with_pdf}
\alias{with_pdf}
\title{Provide a pdf context}
\usage{
with_pdf(
  code,
  ...,
  .file = NULL,
  .width = 7,
  .height = 7,
  .pointsize = 12,
  .envir = parent.frame()
)
}
\arguments{
\item{code}{the plotting code to evaluatie}

\item{...}{graphical parameters}

\item{.width}{figure width, in inches}

\item{.height}{figure height, in inches}

\item{.pointsize}{pointsize}

\item{.envir}{the environment in which to evaluate code}
}
\value{
pdf data
}
\description{
Provide a pdf context
}
