% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{with_png}
\alias{with_png}
\title{Provide a plotting context}
\usage{
with_png(
  code,
  ...,
  .width = 7,
  .height = 7,
  .dpi = 96,
  .envir = parent.frame()
)
}
\arguments{
\item{code}{the plotting code to evaluatie}

\item{...}{graphical parameters}

\item{.width}{figure width, in inches}

\item{.height}{figure height, in inches}

\item{.dpi}{the dots per inch}

\item{.envir}{the environment in which to evaluate code}
}
\value{
png data
}
\description{
Provide a plotting context
}
