% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-wrap.R
\name{wrap_cli_body}
\alias{wrap_cli_body}
\title{Cli wrapper logic}
\usage{
wrap_cli_body(cmd_call, level, .caller_env = caller_env())
}
\arguments{
\item{cmd_call}{a call, the cli operation that we're wrapping}

\item{level}{the logger level to associated with the cmd}

\item{.caller_env}{the caller environment in which to evaluate cmd_call}
}
\description{
wrap_factory uses the body of this function as a building block.  It
does this using 'substitute', replacing the cmd and level parameters
with their known expressions.
}
\details{
Note, the awkward substitute construction:

local_env <- new_environment(list(cmd = cmd))
injection <- lapply(
body(wrap_cli_body),
function(s) do.call(substitute, list(s, local_env))
) \%>\%
as.call()
}
\keyword{internal}
