% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-cli.R
\name{wrap_get_namespace}
\alias{wrap_get_namespace}
\title{Wrapped namespace}
\usage{
wrap_get_namespace(.caller_env)
}
\arguments{
\item{.caller_env}{the caller environment}
}
\value{
the namespace
}
\description{
Wrapped namespace
}
\keyword{internal}
