% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{wrap_imgref_factory}
\alias{wrap_imgref_factory}
\title{Construct a function that will take image data and produce a markdown image reference}
\usage{
wrap_imgref_factory(chunk_id, file_name)
}
\description{
Construct a function that will take image data and produce a markdown image reference
}
