% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-vignette.R
\name{wrap_png}
\alias{wrap_png}
\title{Wrap png_data in an HTML img wrapper}
\usage{
wrap_png(png_data)
}
\arguments{
\item{png_data}{raw png data}
}
\value{
an IMG element
}
\description{
Wrap png_data in an HTML img wrapper
}
