% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-rmd.R
\name{wrapped_doc}
\alias{wrapped_doc}
\title{Wrap an rmarkdown document}
\usage{
wrapped_doc(cmd, args, func)
}
\arguments{
\item{cmd}{unused}

\item{args}{args to be passed to cmd (or func)}

\item{func}{the original function}
}
\description{
This is intended to be used alongside with_monkey_patch
}
\keyword{internal}
