% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pander.R
\name{pander}
\alias{pander}
\alias{pander_return}
\title{wrapped pander functions}
\usage{
pander(x = NULL, ...)

pander_return(...)
}
\description{
These functions are wrapped versions of those in the \href{https://github.com/Rapporter/pander}{pander} package.
This allows us to inject logging metadata into the usual code flow.
}
