% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapped-with-logger.R
\name{wrapped_with_logger}
\alias{wrapped_with_logger}
\title{Wrap a function for logging}
\usage{
wrapped_with_logger(cmd, args, level = NULL)
}
\arguments{
\item{cmd}{a cli function, e.g., cli::cli_alert}

\item{args}{the args to be passed to the cmd via do.call}

\item{level}{the logger level to be associated with the cli method}
}
\description{
Wrap a function for logging
}
