% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapper-cli.R
\name{wrapper_cli}
\alias{wrapper_cli}
\title{Wrapper for a cli method}
\usage{
wrapper_cli(cmd, args, level)
}
\arguments{
\item{cmd}{a cli function, e.g., cli::cli_alert}

\item{args}{the args to be passed to the cmd via do.call}

\item{level}{the logger level to be associated with the cli method}
}
\description{
This includes consideration for cli_inform, cli_warn, and cli_abort which
use an rlang codepath.
}
