% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{app_factory}
\alias{app_factory}
\title{A variant of cli:::clii__inline where we apply a user-specified transformer.}
\usage{
app_factory()
}
\description{
A variant of cli:::clii__inline where we apply a user-specified transformer.
}
