% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/jupyter.R
\name{appender_jupyter}
\alias{appender_jupyter}
\title{An appender for jupyter (with stderr)}
\usage{
appender_jupyter(lines)
}
\description{
In jupyter, stderr will show up on the console, and messages will show up
in the notebook.  In the jupyter context, we also want to display something
that looks like the usual message output.
}
\keyword{internal}
