% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/inject.R
\name{bind_call}
\alias{bind_call}
\title{Bind an injected function}
\usage{
bind_call(ref, cmd_call, ...)
}
\arguments{
\item{ref}{the reference function to bind}

\item{cmd_call}{a cmd call}

\item{...}{additional arguments to add to the local environment}
}
\value{
a "partially evaluated" function
}
\description{
This takes a wrappable function like wrap_cli_body and returns a function with partial
substitutions.  Any remaining args in the wrappable function are available in the returned
method.
}
\keyword{internal}
