% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_wrapped.R
\name{cli_bullets_raw}
\alias{cli_bullets_raw}
\title{Wrapped version of cli::cli_bullets_raw}
\usage{
cli_bullets_raw(text, id = NULL, class = NULL)
}
\arguments{
\item{text}{Character vector of items. See details below on how names
are interpreted.}

\item{id}{Optional id of the \code{div.bullets} element, can be used in themes.}

\item{class}{Optional additional class(es) for the \code{div.bullets} element.}
}
\description{
\code{cli_format_bullets_raw()} is similar to \code{\link[cli:cli_bullets]{cli_bullets()}}, but it does
not perform any inline styling or glue substitutions in the input.
}
