% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-handler.R
\name{cli_nullity_handler}
\alias{cli_nullity_handler}
\title{Let cli methods replace NULL values with sensible substitutions}
\usage{
cli_nullity_handler(msg)
}
\arguments{
\item{msg}{a cli_message}
}
\description{
Let cli methods replace NULL values with sensible substitutions
}
