% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_wrapped.R
\name{cli_output_connection}
\alias{cli_output_connection}
\title{Wrapped version of cli::cli_output_connection}
\usage{
cli_output_connection()
}
\value{
Connection object.
}
\description{
Note that this only refers to the current R process. If the output
is produced in another process, then it is not relevant.
}
