% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_wrapped.R
\name{cli_status_update}
\alias{cli_status_update}
\title{Wrapped version of cli::cli_status_update}
\usage{
cli_status_update(
  id = NULL,
  msg = NULL,
  msg_done = NULL,
  msg_failed = NULL,
  .envir = parent.frame()
)
}
\arguments{
\item{id}{Id of the status bar to update. Defaults to the current
status bar container.}

\item{msg}{Text to update the status bar with. \code{NULL} if you don't want
to change it.}

\item{msg_done}{Updated "done" message. \code{NULL} if you don't want to
change it.}

\item{msg_failed}{Updated "failed" message. \code{NULL} if you don't want to
change it.}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\value{
Id of the status bar container.
}
\description{
\strong{The \verb{cli_status_*()} functions are superseded by
the \code{\link[cli:cli_progress_message]{cli_progress_message()}} and \code{\link[cli:cli_progress_step]{cli_progress_step()}} functions,
because they have a better default behavior.}

Update the status bar
}
\seealso{
This function supports \link[cli:inline-markup]{inline markup}.

The \code{\link[cli:cli_progress_message]{cli_progress_message()}} and \code{\link[cli:cli_progress_step]{cli_progress_step()}}
functions, for a superior API.

Other status bar: 
\code{\link[cli]{cli_process_start}()},
\code{\link[cli]{cli_status_clear}()},
\code{\link[cli]{cli_status}()}

Other functions supporting inline markup: 
\code{\link[cli]{cli_abort}()},
\code{\link[cli]{cli_alert}()},
\code{\link[cli]{cli_blockquote}()},
\code{\link[cli]{cli_bullets_raw}()},
\code{\link[cli]{cli_bullets}()},
\code{\link[cli]{cli_dl}()},
\code{\link[cli]{cli_h1}()},
\code{\link[cli]{cli_li}()},
\code{\link[cli]{cli_ol}()},
\code{\link[cli]{cli_process_start}()},
\code{\link[cli]{cli_progress_along}()},
\code{\link[cli]{cli_progress_bar}()},
\code{\link[cli]{cli_progress_message}()},
\code{\link[cli]{cli_progress_output}()},
\code{\link[cli]{cli_progress_step}()},
\code{\link[cli]{cli_rule}},
\code{\link[cli]{cli_status}()},
\code{\link[cli]{cli_text}()},
\code{\link[cli]{cli_ul}()},
\code{\link[cli]{format_error}()},
\code{\link[cli]{format_inline}()}
}
