% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli_wrapped.R
\name{cli_verbatim}
\alias{cli_verbatim}
\title{Wrapped version of cli::cli_verbatim}
\usage{
cli_verbatim(..., .envir = parent.frame())
}
\arguments{
\item{...}{The text to show, in character vectors. Each element is
printed on a new line.}

\item{.envir}{Environment to evaluate the glue expressions in.}
}
\description{
It is not wrapped, but printed as is. Long lines will overflow.
No glue substitution is performed on verbatim text.
}
