% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_get}
\alias{config_get}
\title{an improved config::get}
\usage{
config_get(
  ...,
  r_config_active = Sys.getenv("R_CONFIG_ACTIVE", "default"),
  file = Sys.getenv("R_CONFIG_FILE", here::here("config.yml")),
  merge.precedence = "override",
  handlers = list(optenv = yaml_handler_optenv, with_env = yaml_handler_with_env)
)
}
\arguments{
\item{...}{passed to purrr::pluck to extract components within the active config}

\item{file}{the yaml file; defaults to config.yml}

\item{config}{the configuration; defaults to R_CONFIG_ACTIVE}
}
\description{
config_get behaves much like config::get.  However, it improves upon its predecessor in the
following ways:
\itemize{
\item handlers for custom YAML tags
\item allows YAML references
}
}
