% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_get}
\alias{config_get}
\title{an improved config::get}
\usage{
config_get(
  ...,
  r_config_active = Sys.getenv("R_CONFIG_ACTIVE", "default"),
  file = config_yml(),
  merge.precedence = "override",
  handlers = list(optenv = yaml_handler_optenv, with_env = yaml_handler_with_env)
)
}
\arguments{
\item{...}{passed to purrr::pluck to extract components within the active config}

\item{r_config_active}{the configuration; defaults to R_CONFIG_ACTIVE}

\item{file}{the config file; defaults to config.yml}

\item{merge.precedence}{merge.precedence; defaults to "override" for YAML references}

\item{handlers}{a list of handlers for YAML tags}
}
\description{
config_get behaves much like config::get.  However, it improves upon its predecessor in the
following ways:
\itemize{
\item handlers for custom YAML tags
\item allows YAML references
}
}
