% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_get}
\alias{config_get}
\title{A variation of config::get that can utilize backreferences}
\usage{
config_get(
  ...,
  r_config_active = Sys.getenv("R_CONFIG_ACTIVE", "default"),
  file = Sys.getenv("R_CONFIG_FILE", here::here("config.yml")),
  merge.precedence = "override",
  handlers = list(optenv = yaml_optenv_handler, with_env = yaml_withenv_handler)
)
}
\arguments{
\item{...}{passed to purrr::pluck}

\item{file}{the yaml file; defaults to config.yml}

\item{config}{the configuration; defaults to R_CONFIG_ACTIVE}
}
\description{
A variation of config::get that can utilize backreferences
}
