% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{config_get}
\alias{config_get}
\title{A version of config::get that can utilize backreferences}
\usage{
config_get(
  ...,
  r_config_active = Sys.getenv("R_CONFIG_ACTIVE", "default"),
  file = Sys.getenv("R_CONFIG_FILE", "config.yml"),
  merge.precedence = "override",
  handlers = NULL
)
}
\arguments{
\item{...}{passed to purrr::pluck}

\item{file}{the yaml file; defaults to config.yml}

\item{config}{the configuration; defaults to R_CONFIG_ACTIVE}
}
\description{
A version of config::get that can utilize backreferences
}
