% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/pubcontext.R
\name{.pubcontext}
\alias{.pubcontext}
\title{Invoke context dependent code}
\usage{
.pubcontext(
  jupyter_expr = NULL,
  knitr_expr = NULL,
  rstudio_expr = NULL,
  interactive_expr = NULL,
  non_interactive_expr = NULL,
  raise = FALSE,
  .envir = parent.frame()
)
}
\description{
Invoke context dependent code
}
