% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdev.R
\name{grdev_adapt}
\alias{grdev_adapt}
\title{subset device args}
\usage{
grdev_adapt(grdev, argset)
}
\arguments{
\item{grdev}{a graphics object acquired from grdev_get}

\item{argset}{a passed argument set}
}
\description{
Update the default graphics device args with argset.  If argset provides
spurious parameters, they will be ignored.
}
\keyword{internal}
