% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdev.R
\name{grdev_get}
\alias{grdev_get}
\title{get grdev information}
\usage{
grdev_get(grdev_name)
}
\arguments{
\item{grdev_name}{the name of the graphics object}
}
\description{
get grdev information
}
\keyword{internal}
