% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdev.R
\name{grdev_resolve}
\alias{grdev_resolve}
\title{resolve file and filename}
\usage{
grdev_resolve(grdev, argset)
}
\arguments{
\item{grdev}{a graphics object acquired from grdev_get}

\item{argset}{a passed argument set}
}
\description{
\verb{file`` and }filename` are grdev specific.  They should both be available downstream and
they should be in agreement.
}
\keyword{internal}
