% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/highlight.R
\name{highlight}
\alias{highlight}
\title{A refactored / shared code highlight function}
\usage{
highlight(code, theme)
}
\arguments{
\item{code}{the code, read from the source file}

\item{theme}{a knitr theme}
}
\description{
A refactored / shared code highlight function
}
\keyword{internal}
