% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-publish.R
\name{ipynb_yaml_extract}
\alias{ipynb_yaml_extract}
\title{Extract YAML from ipynb}
\usage{
ipynb_yaml_extract(ipynb)
}
\arguments{
\item{ipynb}{an ipynb file}
}
\description{
Returns a nested list comprised of merged YAML from raw notebook cells.
}
