% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{logger_reset}
\alias{logger_reset}
\title{Reset the logger logs}
\usage{
logger_reset()
}
\description{
Warning / Experimental:  this accesses private data in the logger package in
an undocumented way
}
\keyword{internal}
