% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metayer-package.R
\docType{package}
\name{metayer-package}
\alias{metayer}
\alias{metayer-package}
\title{metayer: Tooling for package development}
\description{
metayer provides miscellaneous tooling for package development. It currently enhances vignettes written in jupyter lab; allows folding of cli code into logger; adds a hierarchical storage environment; and provides a few utility methods for working with environments.
}
\author{
\strong{Maintainer}: Dustin Lennon \email{dustin.lennon@gmail.com} (\href{https://orcid.org/0009-0002-8558-8765}{ORCID})

}
\keyword{internal}
