% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metayer-package.R
\docType{package}
\name{metayer-package}
\alias{metayer}
\alias{metayer-package}
\title{metayer: An Eclectic Productivity Toolbox}
\description{
Improve the jupyter publication pipeline experience; integrate cli functions into logger; access a global, hierarchical storage container with reference semantics: metayer is an eclectic productivity toolbox intended to help get a new project up and running quickly.
}
\seealso{
Useful links:
\itemize{
  \item \url{https://dlennon.org/metayer/}
}

}
\author{
\strong{Maintainer}: Dustin Lennon \email{dustin.lennon@gmail.com} (\href{https://orcid.org/0009-0002-8558-8765}{ORCID}) [copyright holder]

}
\keyword{internal}
