% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metayer-package.R
\docType{package}
\name{metayer-package}
\alias{metayer}
\alias{metayer-package}
\title{metayer: tools for package development}
\description{
Metayer is a collection of package development tools. It aims to improve the jupyter notebook experience, easing the problem of style drift in later stages of the publication pipeline. It also delivers an integration of the cli and logger packages, imbuing well-known cli operations with logger entry points. The package is intended to be highly configurable, providing extensibility as a reverse dependency.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://localhost}
}

}
\author{
\strong{Maintainer}: Dustin Lennon \email{dustin.lennon@gmail.com} (\href{https://orcid.org/0009-0002-8558-8765}{ORCID})

}
\keyword{internal}
