% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/metayer-package.R
\docType{package}
\name{metayer-package}
\alias{metayer}
\alias{metayer-package}
\title{metayer: tools for package development}
\description{
Metayer is a collection of package development tools. It provides graphic device wrappers that allow uniform figure generation across publishing contexts including Jupyter, knitr, and from the shell. Metayer also integrates the cli and logger packages so that calls to cli methods show up in the logs imbued with package-level context. There is a global, hierarchical storage container that leverages native R environments. Metayer also provides a with-style wrapper that wrapper that facilitates temporary and safe monkey patching of namespaced internals. There are also a few helper methods for inspecting function stack environments associated with dynamic scoping.
}
\seealso{
Useful links:
\itemize{
  \item \url{http://localhost}
}

}
\author{
\strong{Maintainer}: Dustin Lennon \email{dustin.lennon@gmail.com} (\href{https://orcid.org/0009-0002-8558-8765}{ORCID})

}
\keyword{internal}
