% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/logger.R
\name{mocked_log_level_factory}
\alias{mocked_log_level_factory}
\title{a mocked log_level factory}
\usage{
mocked_log_level_factory(logfile = stderr(), envir = parent.frame())
}
\arguments{
\item{logfile}{location for redirecting log data}
}
\description{
a mocked log_level factory
}
