% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-publish.R
\name{preknit}
\alias{preknit}
\title{Knit an rmd file before handing it off to a downstream publishing stage.}
\usage{
preknit(rmd_input, rmd_output, requested_format = c("html", "pdf"))
}
\arguments{
\item{rmd_input}{input Rmd file}

\item{rmd_output}{output Rmd file}

\item{requested_format}{currently, "html" or "pdf"}
}
\description{
Knit an rmd file before handing it off to a downstream publishing stage.
}
