% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-publish.R
\name{prep_articles}
\alias{prep_articles}
\title{prepare rmarkdown articles}
\usage{
prep_articles(..., regexp = "*ipynb", copy_to = NULL)
}
\arguments{
\item{...}{the path for Rmd files, passed to here::here}

\item{regexp}{a regular expression for filtering files}

\item{copy_to}{a destination directory}
}
\description{
Prepare rmarkdown articles for downstream processing.
}
