% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cli-handler.R
\name{preprocess_msg}
\alias{preprocess_msg}
\title{Preprocess a message}
\usage{
preprocess_msg(msg)
}
\arguments{
\item{msg}{a cli message}
}
\value{
the message type
}
\description{
Change NULL values, if metayer.cli_null option is set
}
\keyword{internal}
