% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{pub_ipynb_to_rmd}
\alias{pub_ipynb_to_rmd}
\title{Convert ipynb to knitr-read rmd}
\usage{
pub_ipynb_to_rmd(ipynb, rmd)
}
\arguments{
\item{ipynb}{the ipynb filename}

\item{rmd}{the rmd filename}
}
\description{
Convert ipynb to knitr-read rmd
}
