% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{pub_ipynb_to_rmd}
\alias{pub_ipynb_to_rmd}
\title{convert a notebook to an rmarkdown}
\usage{
pub_ipynb_to_rmd(ipynb_in, rmd_out = NULL)
}
\arguments{
\item{ipynb_in}{the ipynb filename}

\item{rmd_out}{the rmd filename}
}
\description{
Convert a jupyter notebook to a knitr-ready rmarkdown.
}
