% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{pub_rmd_to_html}
\alias{pub_rmd_to_html}
\title{Convert an rmd to an html}
\usage{
pub_rmd_to_html(rmd, html)
}
\arguments{
\item{rmd}{the rmd filename}

\item{html}{the pdf filename}
}
\description{
Convert an rmd to an html
}
