% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/publish.R
\name{pub_rmd_to_pdf}
\alias{pub_rmd_to_pdf}
\title{Convert an rmd to a pdf}
\usage{
pub_rmd_to_pdf(rmd_in, pdf_out = NULL, root_dir = NULL)
}
\arguments{
\item{rmd_in}{the rmd filename}

\item{pdf_out}{the pdf filename}

\item{root_dir}{knitr root.dir}
}
\description{
Convert an rmd to a pdf
}
