% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-test.R
\name{test_sanitize}
\alias{test_sanitize}
\title{Basic test sanitation}
\usage{
test_sanitize(exclusions = c(), envir = parent.frame())
}
\arguments{
\item{exclusions}{a list of object names to exclude from rm.all}

\item{envir}{a local environment}
}
\description{
This:
calls devtools::load_all if not testing
purges the storage
calls rm.all with noted exclusions
sets up a deferred storage purge on exit
sets up a local metayer.verbosity
}
