% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/grdev.R
\name{with_grdev}
\alias{with_grdev}
\title{execute code in a graphics context}
\usage{
with_grdev(
  grdev_name,
  code,
  width = getOption("mty.fig.width", 7),
  height = getOption("mty.fig.height", 7),
  res = getOption("mty.fig.res", 96),
  units = getOption("mty.fig.units", "in"),
  ...,
  par_opts = list(),
  .envir = parent.frame()
)
}
\arguments{
\item{grdev_name}{the name of the graphics device, e.g. "svg" or "png"}

\item{code}{client code block}

\item{width}{width}

\item{height}{height}

\item{res}{resolution}

\item{units}{units of width and height}

\item{...}{passed to dev_args}

\item{par_opts}{graphical parameters, rather than passed through ... in plot}

\item{.envir}{the environment in which to evaluate the code}
}
\description{
Easily execute a code block in a specified graphics context.
}
