% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils-meta.R
\name{wrap_factory}
\alias{wrap_factory}
\title{Wrap an exported function}
\usage{
wrap_factory(pkg, name, level)
}
\arguments{
\item{pkg}{the package name}

\item{name}{the object (function) name}

\item{level}{the logger level}
}
\value{
a wrapped function
}
\description{
Wrap an exported function
}
