% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrapped-factory.R
\name{wrapped_factory}
\alias{wrapped_factory}
\title{Produce a wrapped function}
\usage{
wrapped_factory(name, wrapper, ...)
}
\arguments{
\item{name}{the function name, a string}

\item{wrapper}{the wrapper, a function}

\item{...}{(named) parameters in the execution stack; i.e., private to the wrapper}
}
\description{
A wrapper should have the function signature:
}
\details{
function(cmd, args, ..., .name = NULL)

where ... may, optionally, be replaced with named parameters.

Moreover, at some point in its definition, a wrapper should called the wrapped
function with its args:

do.call(cmd, args)
}
